Fix Map, Switch scene with key and collision portal

+++ 05-ScenceManager/Game.h
@@ -51,7 +51,7 @@ public:
 	void InitKeyboard();
 	void SetKeyHandler(LPKEYEVENTHANDLER handler) { keyHandler = handler; }
 	void Init(HWND hWnd);
-	void Draw(float x, float y, LPDIRECT3DTEXTURE9 texture, int left, int top, int right, int bottom, int alpha = 255, int isFLipImage = 0);
+	void Draw(int x, int y, LPDIRECT3DTEXTURE9 texture, int left, int top, int right, int bottom, int alpha = 255, int isFLipImage = 0);
 
 	int IsKeyDown(int KeyCode);
 	void ProcessKeyboard();
